refactor(MANY): Fix imports for pybuilder closes #40
