+++ docs/source/release1_1_1.rst
@@ -1,0 +1,66 @@
+.. _Pybuilder: http://pybuilder.github.io/
+.. _Google_Cloud: https://cloud.google.com/
+.. _release1_1_1:
+
+############################
+Notas de Lançamento V. 1.1.1
+############################
+
+*SuperPython*
+
+Milestone
+=========
+
+hidenseek - Suporte a importação cruzada
+
+Aspectos do Lançamento
+======================
+
+Destaques dos Aspectos
+**********************
+
+Este ambiente permite a edição de códigos e importação do módulo de um outro projeto.
+
+Aspecto #1
+**********
+
+o nome __name__ é atribuído "__main__" no programa principal.
+
+Melhoramentos
+=============
+
+Versão para uso do Google Application Engine e PyBuilder.
+
+Melhoramento #1
+***************
+
+Foi melhorada a configuração para uso do construtor de aplicativos PyBuilder_. A configuração *build.py*
+define os aspectos necessários para verificar testes, cobertura, cabeçalhos e detalhes para deployment.
+
+Consertos
+=========
+
+Conserto #01
+************
+
+o nome __name__ é atribuído "__main__" no programa principal.
+
+Conserto #02
+************
+
+A carga do módulo principal é feita por AJAX, evitando o congelamento do HTML.
+
+Questões e Problemas Conhecidos
+===============================
+
+A funcionalidade ainda é muito simples, requer melhorias.
+
+Uma nova versão deve suportar o monitoramento  da atividade dos alunos.
+
+Lançamentos Anteriores e Posteriores
+====================================
+
+Próximo Lançamento: A ser definido :ref:`Lançamento 1.1.2 <release1_1_1>`
+
+Lançamento Anterior: Julho 2015 :ref:`Lançamento 0.1.0 <release0_1_0>`
+
