+++ docs/source/tests.rst
@@ -1,8 +1,8 @@
 .. _testes_superpython:
 
-################################
-SuperPython - Módulos Principais
-################################
+##############################
+SuperPython - Módulos de Teste
+##############################
 
 .. automodule:: superpython_tests
     :members:
@@ -19,3 +19,19 @@ SuperPython - Módulos Principais
 .. note::
 
    Módulo principal de testes unitários.
+
+
+.. automodule:: spyweb_tests
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :platform: Web
+    :synopsis: Unidade de Teste da funcionalide Web.
+
+
+.. seealso::
+
+   Module :mod:`spyweb_tests`
+
+.. note::
+   Unidade de Teste da funcionalide Web.
