+++ requirements.txt
@@ -1,3 +1,4 @@
 bottle==0.11.6
 pybuilder==0.10.36
+minimock==1.2.8
